requirements.txt
undetected-chromedriver
selenium
argparse
python-dotenv